Move browser AG-UI encoding into the public agent surface
The hosted runtime formatter and downstream agent convergence both needed a framework-owned browser AG-UI encoder. This extracts the runtime-event mapping into a public agent module, reuses it from the internal hosted formatter, and documents the supported seam so downstream services can stop depending on private transport ownership.

Constraint: veryfront-agent should not depend on internal-agents transport modules
Rejected: Export internal-agents/ag-ui-sse directly | would leak a private hosted-run transport module instead of publishing a clean public seam
Confidence: high
Scope-risk: moderate
Directive: Keep browser/public AG-UI event ownership in the public agent surface; do not reintroduce downstream-only mappings without checking this encoder first
Tested: deno test --no-check --allow-all src/agent/ag-ui-browser-encoder.test.ts src/internal-agents/ag-ui-sse.test.ts src/agent/ag-ui-handler.test.ts; deno lint src/agent/ag-ui-browser-encoder.ts src/agent/ag-ui-browser-encoder.test.ts src/internal-agents/ag-ui-sse.ts src/agent/index.ts; deno check src/agent/ag-ui-browser-encoder.ts src/agent/ag-ui-browser-encoder.test.ts src/internal-agents/ag-ui-sse.ts src/agent/index.ts; deno task verify
Not-tested: downstream veryfront-agent adoption against a published package containing this seam
Related: veryfront-code#973
